Create Base Solution
This commit creates the base solution for binary
encoding.The main function,
decimal_to_binary(num_string), converts a decimal
number string into IEEE-754 64 bit binary. The
main method generates an interface in the
terminal that allows the user to convert decimals
to binary.

Issue:
This function is fully functional for integers
and mostly functional for fractions. The last
few digits of the mantissa are not always accurate
due to floating point error. To fix this, I may
switch from using floats to decimals.

To-Do:
- switch from floats to decimals (for fractions)
- check/consider more edge cases
- exception handling
In short, make the solution more robust.