Add Figures Class and Equation Creation
This commit introduces a `Figure` class to streamline the process of
generating and saving plots. Plotting function calls have been
refactored in the main script to utilize this new class.

Additionally, a function to generate the Lagrange polynomial in LaTeX
format has been implemented. Generated equations are written to txt
files in the main method.